pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lab-notebook"
version = "0.1.0"
description = "A web-based application for viewing and organizing laboratory experiment plots chronologically"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "flask>=3.0.2",
    "python-dateutil>=2.8.2",
    "numpy>=1.26.4",
    "h5py>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["lab_notebook"]

[tool.hatch.build]
include = [
    "lab_notebook/**/*.py",
    "lab_notebook/templates/**/*.html",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 